backend/requirements.txt
# Core dependencies
beautifulsoup4==4.12.3
requests==2.31.0
playwright==1.42.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.1

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1  # MongoDB Python driver
mongoengine==0.27.0  # ODM for MongoDB

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.2
loguru==0.7.2 